charge_controller/requirements.txt
distlib==0.3.9
filelock==3.18.0
minimalmodbus==2.1.1
platformdirs==4.3.8
prometheus-client==0.1.0
pyserial==3.4
pytz==2025.2
renogymodbus==0.0.23
retrying==1.3.4
RPi.bme280==0.2.4
RPi.GPIO==0.7.1
six==1.17.0
smbus2==0.5.0
virtualenv==20.31.2
